Change impedance type from Float to Double
openScale database expects Impedance to be a Double

+++ android_app/app/src/main/java/com/health/openscale/core/bluetooth/scales/SenssunIFB7Handler.kt
@@ -89,7 +89,7 @@ class SenssunIFB7Handler : ScaleDeviceHandler() {
             val finalMeasurement = ScaleMeasurement().apply {
                 weight = weightKg
                 if (impedanceRaw > 0) {
-                    impedance = impedanceRaw.toFloat()
+                    impedance = impedanceRaw.toDouble()
                 }
             }
             publish(finalMeasurement)
